Don't know how to do fast 16 bit loops in 65c02 ASM...

+++ src/fill.h
@@ -1,5 +1,5 @@
 #ifndef FILL_H
 #define FILL_H
 #include <stdint.h>
-extern void __fastcall__ fill_vera(uint16_t count, uint8_t data);
+extern void __fastcall__ fill_vera(uint8_t count, uint8_t data);
 #endif
